A spike about setting non-zero amount description required
This approach has the timing problem that validation (at least
the required-validation of the description field) should be
carried out only in the render phase, outside the normal
validation phase to get the validation result show right away.
Probably this needs to be rather done inside the field update loop.